+++ programs/ralli-bet/programs/ralli-bet/src/instructions/resolve_line.rs
@@ -17,7 +17,7 @@ pub struct ResolveLine<'info> {
 }
 
 impl<'info> ResolveLine<'info> {
-    pub fn resolve_line(&mut self, result: Direction, actual_value: i64) -> Result<()> {
+    pub fn resolve_line(&mut self, result: Direction, actual_value: f64) -> Result<()> {
         let admin = &self.admin;
         let line = &mut self.line;
